+++ packages/backend-output-storage/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @aws-amplify/backend-output-storage
 
+## 1.1.1
+
+### Patch Changes
+
+- a65371c: upgrade aws-cdk and aws-cdk-lib to ^2.152.0
+- Updated dependencies [4cce19f]
+  - @aws-amplify/platform-core@1.0.6
+
 ## 1.1.0
 
 ### Minor Changes